refactor: refactored post page to use components
